refactor: admin 용어 상세 정보 form 컴포넌트 분리
